requirements.txt
flask
flask-restful
python-dotenv
google-generativeai